+++ src/hydra/dataProvider.ts
@@ -257,7 +257,7 @@ function dataProvider(
       const values = Object.values(hydraData);
       const containFile = (element: unknown): boolean =>
         Array.isArray(element)
-          ? element.every((value) => containFile(value))
+          ? element.length > 0 && element.every((value) => containFile(value))
           : isPlainObject(element) &&
             Object.values(element as Record<string, unknown>).some(
               (value) => value instanceof File,